requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.3.0
scipy==1.10.0
tensorflow==2.12.0
torch==2.0.1
torchvision==0.15.2
transformers==4.39.3
opencv-python==4.8.0.74
pillow==10.4.0
matplotlib==3.7.1
h5py==3.8.0
Flask==2.3.2
Flask-Login==0.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.0.0
Flask-Migrate==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
joblib==1.2.0